requirements.txt
fastapi
uvicorn
gunicorn
requests
python-dotenv
pydantic
sentence-transformers
pinecone[grpc]